WIP: updated test_id_cooccurrence_v2.ipynb.
Added a function resolve_2way_conflicts with the attempt to deal with same feature appearing in multiple 2-way overlaps.
While the function does what it was intended to do, the results reveal such approach is logically incorrect.

For example, if an AR overlaps with two MCSs and an ETC, with one of the MCSs having the largest overlap fraction (with respect to the AR area), then only that MCS is retained, while the other MCS and ETC are removed. This results in many 2-way overlap features being removed, which is not the intended purpose.

A different approach is needed.